feat: move selection buttons into their own functions
